+++ Changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v0.4.0
+### Added
+- Class BaseMultiStepValueObject
+- Class BaseValueObject
+
 ## v0.3.0
 ### Added
 - removeAll for session data